Simplify utility usage. Remove symbol support (use state instead).
