+++ packages/skatejs/src/name.js
@@ -1,6 +1,14 @@
 // @flow
 
-import { dashCase } from './util';
+export function dashCase<T>(str: T): T {
+  return typeof str === 'string'
+    ? str.split(/([_A-Z])/).reduce((one, two, idx) => {
+        const dash = !one || idx % 2 === 0 ? '' : '-';
+        two = two === '_' ? '' : two;
+        return `${one}${dash}${two.toLowerCase()}`;
+      })
+    : str;
+}
 
 function format(prefix: string, suffix: number): string {
   return (
